Add tests and seeding, enable compiler warnings
- -Wall -Wextra on all targets; fix the existing warnings (unused
  variables, initializer order, sign compare, && within ||, class/struct
  tag mismatch).
- seed_rand(seed); dev_seed_rand() returns the seed it chose.  rand_gauss's
  cached deviate moves into the generator state so reseeding resets it
  (otherwise two runs in one process could differ).
- hopper -S <seed> repeats a run exactly; the seed is printed and saved as
  seed.txt in the run directory.
- Tests (ctest), plain executables using tests/check.h:
  - neatTests: genome save/load and clone round trips, compat properties,
    mating invariants, networks vs hand calculations, innovation sharing,
    seeded RNG, seeded GA gives the same history twice
  - creatureConfig: minimal config builds, bad names/types/missing fields
    fail cleanly, shipped configs build
  - determinism: two seeded hopper runs write identical genomes,
    snapshots and species.csv (replaces testMult, now removed)

+++ NEAT/GeneticAlgorithm.cpp
@@ -480,7 +480,7 @@ double GeneticAlgorithm::nextGeneration() {
 
 void GeneticAlgorithm::printPopulation() const {
     cout<<"Population size: "<<population.size()<<endl;
-    for (int i = 0; i<population.size(); ++i) {
+    for (size_t i = 0; i<population.size(); ++i) {
 	cout<<"Member "<<i<<":"<<endl;
 	population[i]->printDescription("  ");
     }
